fix: type

+++ executor/simple_executor.go
@@ -248,7 +248,7 @@ func (e *InsertExecutor) Execute(pl planner.InsertPlan) (*ResultSet, error) {
 
 	if err := btree.Insert(&storage.StringItem{
 		Value:  pl.PKValue,
-		PageId: page.Id,
+		PageId: uint64(page.Id),
 	}); err != nil {
 		return nil, err
 	}
